errorSigHashToName
