+++ packages/sdk/src/client.ts
@@ -33,8 +33,9 @@ import {
   NULL_BYTES32,
   MINT_GAS_LIMIT_MULTIPLIER,
   MINT_FALLBACK_GAS_LIMIT,
+  errorSigHashToName,
 } from './utils/constants'
-import { getLazyOption, getSaltAsBytes32, validateAddress, scaleAmount, getErrorSelectors } from './utils/helpers'
+import { getLazyOption, getSaltAsBytes32, validateAddress, scaleAmount } from './utils/helpers'
 import { LazyPromise } from './utils/promise'
 
 import type {
@@ -682,11 +683,11 @@ export function SoundClient({
         tx.blockNumber,
       )
 
-      // If this is a failed transaction, the first 4 bytes of the response will be the custom error selector (hash of its signature)
+      // If this is a failed transaction, the first 4 bytes of the response
+      // will be the custom error selector (hash of its signature)
       const firstFourBytes = response.slice(0, 10)
-      const errorSelectors = getErrorSelectors(tx.data)
 
-      return errorSelectors[firstFourBytes] ?? 'Unable to parse error'
+      return errorSigHashToName[firstFourBytes] ?? 'Unable to parse error'
     } catch (err) {
       console.error(err)
     }
